dbaas: read-replica subcommands (#834)
# Description
<!--
* Prefix: the title with the component name being changed. Add a short
and self describing sentence to ease the review
* Please add a few lines providing context and describing the change
* Please self comment changes whenever applicable to help with the
review process
* Please keep the checklist as part of the PR. Tick what applies to this
change.
-->


[ADR](https://github.com/exoscale/dbaas-api/blob/master/doc/adr/011-read-replica-tooling.md)

```
 exo dbaas read-replica
Manage Database Service read replicas

Usage:
  exo dbaas read-replica [command]

Available Commands:
  create      Create a Database Service read replica
  list        List Database Service read replicas
  promote     Promote a Database Service read replica to a standalone primary
  show        Show Database Service read replica details
```

```
exo dbaas read-replica create --help
Create a read replica for an existing PostgreSQL or MySQL/MariaDB primary Database Service. The replica can be created in a different zone than the primary.

Usage:
  exo dbaas read-replica create READ-REPLICA-NAME [flags]

Aliases:
  create, add

Flags:
  -h, --help                     help for create
      --plan string              subscription plan
  -z, --replica-zone string      zone where the replica will be created
      --source-service string    name of the primary service
      --termination-protection   enable Database Service termination protection; set --termination-protection=false to disable (default true)
```

```
exo dbaas read-replica show --help
Show details of a specific Database Service read replica.

Usage:
  exo dbaas read-replica show REPLICA-NAME [flags]

Aliases:
  show, get

Flags:
  -h, --help          help for show
  -z, --zone string   Database Service zone
  
  ```
  
  ```
  exo  dbaas read-replica list --help
List all read replicas of a primary Database Service across all zones.

Usage:
  exo dbaas read-replica list SERVICE-NAME [flags]

Aliases:
  list, ls
  ``` 
  
  ```
  exo dbaas read-replica promote --help
Promote a read replica to a standalone primary Database Service. This breaks the replication link.

Usage:
  exo dbaas read-replica promote REPLICA-NAME [flags]

Flags:
  -f, --force         don't prompt for confirmation
  -h, --help          help for promote
  -z, --zone string   Database Service zone
  ```

## Checklist
(For exoscale contributors)

* [x] Changelog updated (under *Unreleased* block, and add the Pull Request #number for each bit you add to the `CHANGELOG.md`)
* [x] Testing

## Testing
Tested in pre-prod with locally built binary.
<!--
Describe the tests you did
-->

---------

+++ cmd/dbaas/dbaas_delete.go
@@ -3,6 +3,7 @@ package dbaas
 import (
 	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/spf13/cobra"
 
@@ -47,6 +48,18 @@ func (c *dbaasServiceDeleteCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		if !utils.AskQuestion(ctx, fmt.Sprintf("Are you sure you want to delete Database Service %q?", c.Name)) {
 			return nil
 		}
+	}
+
+	svc, err := dbaasGetV3(ctx, c.Name, c.Zone)
+	if err != nil {
+		return err
+	}
+
+	readReplicaNames := dbaasActiveReadReplicaNamesForPrimary(svc)
+	if len(readReplicaNames) > 0 {
+		return fmt.Errorf(
+			"cannot delete Database Service %s with active read replica(s): %s. Delete or promote the read replica(s) first",
+			c.Name, strings.Join(readReplicaNames, ", "))
 	}
 
 	op, err := client.DeleteDBAASService(ctx, c.Name)
